Keep release-channel test behavioral
